==> P3_FilterUnwanted.ipynb <==
# Part 3: Filter compounds for PAINS and other unwanted substructures

Unwanted side effects are not desired for promising compounds. In this part will compounds with known unwanted substructures filtered out of the dataset. 

Therefore Pan-assay interference compounds (PAINS) will be filtered out of the dataset, as these structures bind unspecifically, which can result in inwanted side effects. Filtering those PAINS can be done by filtering on the substructres that are desribed by Baell et al. (2010). Moreover, Brenk et al. (2008) provided a list of other unwatned substructures, which will be used for futher substructure filtering. 

Import required libraries

from pathlib import Path

import pandas as pd
from tqdm.auto import tqdm
from rdkit import Chem
from rdkit.Chem import PandasTools
from rdkit.Chem.FilterCatalog import FilterCatalog, FilterCatalogParams

Define paths to this notebook

HERE = Path(_dh[-1])
DATA = HERE / "data"

Read data from Part 2

data = pd.read_csv(DATA/"BACE_compounds_lipinski.csv",
    index_col=0,
)
data

Drop the columns molecular weight, n_hbd, n_hba, logp as they are not needed anymore. They were only used for to check if it fullfill the RO5. 

print("Dataframe shape:", data.shape)
data.drop(columns=["molecular_weight", "n_hbd", "n_hba", "logp"], inplace=True)
data.head()

Add molecule column

PandasTools.AddMoleculeColumnToFrame(data, smilesCol="smiles")
data.head()

## Filter for PAINS

Remove PAINS (pan-assay interference compounds) because of high binding affinity to off-targets.

params = FilterCatalogParams()
params.AddCatalog(FilterCatalogParams.FilterCatalogs.PAINS)  # add PAINS from catalogue to filter
catalog = FilterCatalog(params)

Search for PAINS in the dataset and keep the molecules without PAINS

matches = []
clean = []
for index, row in tqdm(data.iterrows(), total=data.shape[0]):
    molecule = Chem.MolFromSmiles(row.smiles)
    entry = catalog.GetFirstMatch(molecule)  # Get the first matching PAINS
    if entry is not None:
        # store PAINS information
        matches.append(
            {
                "chembl_id": row.molecule_chembl_id,
                "rdkit_molecule": molecule,
                "pains": entry.GetDescription().capitalize(),
            }
        )
    else:
        # collect indices of molecules without PAINS
        clean.append(index)

matches = pd.DataFrame(matches)
data = data.loc[clean]  # keep molecules without PAINS

print(f"Number of compounds with PAINS: {len(matches)}")
print(f"Number of compounds without PAINS: {len(data)}")
print(f"percentage of compounds with PAINS: {round(len(matches)/len(data)*100,2)}%")


The compounds in the dataset are filtered for PAINS in order to avoid unwanted side-effects. 133 (2.03%) compounds contained PAINS and were therefore removed from the dataset. 

## Filter for other unwanted substructures

Read file with the unwanted substructures obtained from article (Chem. Med. Chem. (2008), 3, 535-44)

In the table below are the unwanted substructures shown

substructures = pd.read_csv(DATA/"unwanted_substructures.csv", sep="\s+")
substructures["rdkit_molecule"] = substructures.smarts.apply(Chem.MolFromSmarts)
print("Number of unwanted substructures in collection:", len(substructures))
display(substructures) # Show the substructures that are unwanted 

Search for these unwanted substructures in the dataset and keep the molecules without unwanted substructures

matches = []
clean = []
for index, row in tqdm(data.iterrows(), total=data.shape[0]):
    molecule = Chem.MolFromSmiles(row.smiles)
    match = False
    for _, substructure in substructures.iterrows():
        if molecule.HasSubstructMatch(substructure.rdkit_molecule):
            matches.append(
                {
                    "chembl_id": row.molecule_chembl_id,
                    "rdkit_molecule": molecule,
                    "substructure": substructure.rdkit_molecule,
                    "substructure_name": substructure["name"],
                }
            )
            match = True
    if not match:
        clean.append(index)

matches = pd.DataFrame(matches)
data = data.loc[clean]

print(f"Number of found unwanted substructure: {len(matches)}")
print(f"Number of compounds without unwanted substructure: {len(data)}")
print(f"percentage of compounds with unwanted substructure: {round(len(matches)/len(data)*100,2)}%")


The current dataset contained 2319 (48.08%) compounds with unwanted substures from the list in the article of Brenk et al. (2008). These are removed, resulting in a dataset of 2319 remaining compounds.

List with frequenction that the unwanted substructures are present in the data 

groups = matches.groupby("substructure_name")
group_frequencies = groups.size()
group_frequencies.sort_values(ascending=False, inplace=True)
group_frequencies

The most common of those unwanted structures were imines, triple-bonds and 2-haloPyridine

Save filtered data to a csv file 

data = data.drop("ROMol", axis=1)
data.to_csv(DATA/"BACE_compounds_part3.csv")
data.head()

This filtering result in a filtered dataset of 4823 compounds. This dataset will be used to draw a scaffold. This data will also be used to train and validate the machine learning models.
