==> P9.1_Regressionmodel.ipynb <==
# Part 9: Random Forest regression model

The potency of the novel generated compounds for the BACE1 receptor will be determined with machine learning models. In this project will two different models be compared; a random forest regression model and a neuron network model. These models will be trained and validated with the filtered ChEMBL bioactivity data from part 3. In this part will the random forest regression model be trained and validated, followed by using the model to predict the potency of the novel compounds.

Import required libraries

from pathlib import Path
from warnings import filterwarnings
import time

import pandas as pd
import numpy as np
from sklearn import svm, metrics, clone
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import RandomForestRegressor
from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import KFold, train_test_split
from sklearn.metrics import auc, accuracy_score, recall_score
from sklearn.metrics import roc_curve, roc_auc_score
import matplotlib.pyplot as plt
from rdkit import Chem
from rdkit.Chem import MACCSkeys
from rdkit.Chem.AllChem import GetMorganFingerprintAsBitVect
from rdkit.Chem import PandasTools

import sys

from teachopencadd.utils import seed_everything
SEED = 22
seed_everything(SEED)


Define paths

HERE = Path(_dh[-1])
DATA = HERE / "data"

Read filtered ChEMBL bioactivity data from part 3

chembl_df = pd.read_csv(DATA/"BACE_compounds.csv",
    index_col=0,
)

print("Shape of dataframe : ", chembl_df.shape)
chembl_df.head()

Keep only the needed columns

chembl_df = chembl_df[["molecule_chembl_id", "smiles", "pIC50"]]
chembl_df.head()

Add column for fingerprints

compound_df = chembl_df.copy()

Set the number of folds

N_FOLDS = 5

Fingerprints. The structures of the compounds will be transferred from SMILES notation to Morgan Fingerprints with a radius of 3, to explicitly describe the features of the compounds.

Define function to convert the smiles into fingerprints

def smiles_to_fp(smiles, method="maccs", n_bits=2048):
    """
    Encode a molecule from a SMILES string into a fingerprint.

    Parameters
    ----------
    smiles : str
        The SMILES string defining the molecule.

    method : str
        The type of fingerprint to use. Default is MACCS keys.

    n_bits : int
        The length of the fingerprint.

    Returns
    -------
    array
        The fingerprint array.

    """

    # convert smiles to RDKit mol object
    mol = Chem.MolFromSmiles(smiles)

    if method == "maccs":
        return np.array(MACCSkeys.GenMACCSKeys(mol))
    if method == "morgan2":
        return np.array(GetMorganFingerprintAsBitVect(mol, 2, nBits=n_bits))
    if method == "morgan3":
        return np.array(GetMorganFingerprintAsBitVect(mol, 3, nBits=n_bits))
    else:
        # NBVAL_CHECK_OUTPUT
        print(f"Warning: Wrong method specified: {method}. Default will be used instead.")
        return np.array(MACCSkeys.GenMACCSKeys(mol))

Use Morgan fingerprint with radius 3

compound_df["fp"] = compound_df["smiles"].apply(smiles_to_fp, args=("morgan3",))
compound_df.head()

Define function for machine learning model training and validation in a cross-validation loop

def crossvalidation_reg(ml_model, df, n_folds=5, verbose=False):
    """
    Machine learning model training and validation in a cross-validation loop.

    Parameters
    ----------
    ml_model: sklearn model object
        The machine learning model to train.
    df: pd.DataFrame
        Data set with SMILES and their associated activity labels.
    n_folds: int, optional
        Number of folds for cross-validation.
    verbose: bool, optional
        Performance measures are printed.

    Returns
    -------
    None

    """
    t0 = time.time()
    # Shuffle the indices for the k-fold cross-validation
    kf = KFold(n_splits=n_folds, shuffle=True)#, random_state=SEED)

    # Results for each of the cross-validation folds
    MAE_per_fold = []
    RMSE_per_fold = []

    # Loop over the folds
    for train_index, test_index in kf.split(df):
        # clone model -- we want a fresh copy per fold!
        fold_model = clone(ml_model)
        # Training

        # Convert the fingerprint and the label to a list
        train_x = df.iloc[train_index].fp.tolist()
        train_y = df.iloc[train_index].pIC50.tolist()

        # Fit the model
        fold_model.fit(train_x, train_y)

        # Testing

        # Convert the fingerprint and the label to a list
        test_x = df.iloc[test_index].fp.tolist()
        test_y = df.iloc[test_index].pIC50.tolist()

        test_results = fold_model.predict(test_x)
        # Prediction probability on test set
        from sklearn import metrics

        MAE_per_fold.append(metrics.mean_absolute_error(test_y, test_results))
        #print('Mean Squared Error (MSE):', metrics.mean_squared_error(test_y, test_results))
        RMSE_per_fold.append(np.sqrt(metrics.mean_squared_error(test_y, test_results)))

        #from sklearn.metrics import matthews_corrcoef


        #print(matthews_corrcoef(test_y, test_results))
        #mape = np.mean(np.abs((gt - pred) / np.abs(gt)))
        #print('Mean Absolute Percentage Error (MAPE):', round(mape * 100, 2))
        #print('Accuracy:', round(100*(1 - mape), 2))
    return(MAE_per_fold,RMSE_per_fold,fold_model)

Random forest regressor model

The benefit of a regression model as opposed to a classification model is that it can not only predict if the new compound is active or inactive, but also the pIC50 value of the compound.

models = [{"label": "Model_RF_reg", "model": RandomForestRegressor}]

Train model with RandomForestRegressor and show validation scores

regressor = RandomForestRegressor()
MAE, RMSE,trained_model = crossvalidation_reg(regressor, compound_df, n_folds=N_FOLDS)

print(
f"Mean Absolute Error (MAE): {np.mean(MAE):.2f} \t"
f"and std : {np.std(MAE):.2f} \n"
f"Root Mean Square Error (RMSE): {np.mean(RMSE):.2f} \t"
f"and std : {np.std(RMSE):.2f} \n"
)


compound_df.head()

To validate the performance of the model, the data was split into a train and test set. In this case this is done in a 5-fold cross-validation. Separate models are trained, each with one of these folds as test data, and all other folds as training data.
We then considered the overall performance. The errors between the predicted data and obtained experimental data are determined with a Mean Absolute Error (MAE) and a Root Mean Square Error (RMSE).


print(trained_model)

This trained model will below be used to predict the pIC50 values of the novel, drug-like, generated compounds of part 8.

### Generated Compounds: predict pIC50 with the trained random forest regression model

Read the filtered, generated compound data of part 8

novelcompound_df = pd.read_csv(
   DATA / "generated_part8.csv", index_col=0,
)

print("Shape of dataframe : ", novelcompound_df.shape)
novelcompound_df.head()

List with SMILES

novelcompound_df_smiles = novelcompound_df.SMILES.tolist()

Convert SMILES to morgan fingerprints with radius 3. Use these fingerprints to predict the pIC50 of these compounds with the trained random forest regression model. 

fps = []

for smiles in novelcompound_df_smiles: 
    fp = smiles_to_fp(smiles,'morgan3')
    fps.append(fp)

predictions = trained_model.predict(fps)

Show results

novelcompound_df_smiles_results = pd.DataFrame({
"SMILES" : novelcompound_df_smiles,
"predicted_pIC50" : predictions
})

novelcompound_df_smiles_results_m = pd.merge(
    novelcompound_df_smiles_results,
    novelcompound_df,
    on = "SMILES",
)

PandasTools.AddMoleculeColumnToFrame(novelcompound_df_smiles_results_m, smilesCol="SMILES")
novelcompound_df_smiles_results_m.sort_values(by="predicted_pIC50", ascending = False, inplace = True)
novelcompound_df_smiles_results_m.reset_index(drop=True, inplace=True)
novelcompound_df_smiles_results_m.head(10)

As can been seen in the table above, the all top 10 molecules fulfil Lipinski's rule of 5, which could mean that they are oral bioavailable. The molecule with the highest predicted pIC50 value is N-(3-(5-amino-3,6,6-trimethyl-1,1-dioxido-3,6-dihydro-2H-1,4-thiazin-3-yl)-4-fluorophenyl)-5-methoxypicolinamide. This molecule does not contain the scaffold of the largest cluster. However, it has the common substructure of the high potent inhibitors (threshold 0.7, with and without ring match). The predicted pIC50 of this compound is 11.37, which is quite high. Moreover, the molecule is not known in the PubChem database. Therefore, this can be a promising, novel inhibitor of BACE1. 

Save results of the novel compounds

output_df = novelcompound_df_smiles_results_m.drop("ROMol", axis = 1)
output_df.to_csv(DATA/"nieuwmoleculenlijst_voorspellingen.csv")
output_df.head()
